+++ packages/cli/src/plugins/transforms/transform.assets.js
